move getCurrentFee function to shared place
