+++ extension/src/popup/helpers/fees.ts
@@ -1,0 +1,13 @@
+import { BASE_FEE } from "stellar-sdk";
+import { stroopToXlm } from "helpers/stellar";
+
+export const getCurrentTransactionFee = ({
+  currentTransactionFee,
+  fallbackTransactionFee,
+}: {
+  currentTransactionFee?: string;
+  fallbackTransactionFee?: string;
+}) =>
+  currentTransactionFee ||
+  fallbackTransactionFee ||
+  stroopToXlm(BASE_FEE).toString();
